Remove inject methods as far as possible
