+++ Classes/Controller/EventController.php
@@ -11,15 +11,33 @@ declare(strict_types=1);
 
 namespace JWeiland\Events2\Controller;
 
+use JWeiland\Events2\Configuration\ExtConf;
 use JWeiland\Events2\Domain\Model\Event;
+use JWeiland\Events2\Domain\Model\Link;
 use JWeiland\Events2\Domain\Model\Search;
+use JWeiland\Events2\Domain\Repository\CategoryRepository;
+use JWeiland\Events2\Domain\Repository\DayRepository;
+use JWeiland\Events2\Domain\Repository\EventRepository;
+use JWeiland\Events2\Domain\Repository\LinkRepository;
+use JWeiland\Events2\Domain\Repository\LocationRepository;
+use JWeiland\Events2\Domain\Repository\UserRepository;
+use JWeiland\Events2\Helper\HiddenObjectHelper;
 use JWeiland\Events2\Property\TypeConverter\UploadMultipleFilesConverter;
 use JWeiland\Events2\Service\DayRelationService;
 use JWeiland\Events2\Utility\CacheUtility;
+use TYPO3\CMS\Core\Mail\MailMessage;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Utility\MathUtility;
+use TYPO3\CMS\Extbase\Persistence\PersistenceManagerInterface;
 use TYPO3\CMS\Extbase\Property\TypeConverter\DateTimeConverter;
 use TYPO3\CMS\Extbase\Property\TypeConverter\PersistentObjectConverter;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
+use TYPO3\CMS\Extbase\Validation\Validator\ConjunctionValidator;
+use TYPO3\CMS\Extbase\Validation\Validator\GenericObjectValidator;
+use TYPO3\CMS\Extbase\Validation\Validator\NotEmptyValidator;
+use TYPO3\CMS\Extbase\Validation\Validator\RegularExpressionValidator;
+use TYPO3\CMS\Extbase\Validation\Validator\ValidatorInterface;
+use TYPO3\CMS\Extbase\Validation\ValidatorResolver;
 
 /*
  * The EventController contains actions to search, create, activate, update and delete events.
@@ -27,6 +45,78 @@ use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
  */
 class EventController extends AbstractController
 {
+    /**
+     * @var DayRepository
+     */
+    protected $dayRepository;
+
+    /**
+     * @var EventRepository
+     */
+    protected $eventRepository;
+
+    /**
+     * @var CategoryRepository
+     */
+    protected $categoryRepository;
+
+    /**
+     * @var LocationRepository
+     */
+    protected $locationRepository;
+
+    /**
+     * @var LinkRepository
+     */
+    protected $linkRepository;
+
+    /**
+     * @var UserRepository
+     */
+    protected $userRepository;
+
+    /**
+     * @var PersistenceManagerInterface
+     */
+    protected $persistenceManager;
+
+    /**
+     * @var MailMessage
+     */
+    protected $mail;
+
+    /**
+     * @var ExtConf
+     */
+    protected $extConf;
+
+    public function __construct(
+        DayRepository $dayRepository,
+        EventRepository $eventRepository,
+        CategoryRepository $categoryRepository,
+        LocationRepository $locationRepository,
+        LinkRepository $linkRepository,
+        UserRepository $userRepository,
+        PersistenceManagerInterface $persistenceManager,
+        MailMessage $mail,
+        ExtConf $extConf
+    ) {
+        $this->dayRepository = $dayRepository;
+        $this->eventRepository = $eventRepository;
+        $this->categoryRepository = $categoryRepository;
+        $this->locationRepository = $locationRepository;
+        $this->linkRepository = $linkRepository;
+        $this->userRepository = $userRepository;
+        $this->persistenceManager = $persistenceManager;
+        $this->mail = $mail;
+        $this->extConf = $extConf;
+    }
+
+    public function initializeObject()
+    {
+        $this->dayRepository->setSettings($this->settings);
+    }
+
     /**
      * We have a self-build form based on method GET.
      * That's why we have to manually allow some form-elements.
@@ -135,7 +225,7 @@ class EventController extends AbstractController
      */
     public function editAction(int $event): void
     {
-        $eventObject = $this->eventRepository->findHiddenEntry($event);
+        $eventObject = $this->eventRepository->findHiddenObject($event);
         $categories = $this->categoryRepository->getCategories(
             $this->settings['selectableCategoriesForNewEvents']
         );
@@ -155,7 +245,12 @@ class EventController extends AbstractController
      */
     public function initializeUpdateAction(): void
     {
-        $this->registerEventFromRequest();
+        $hiddenObjectHelper = $this->objectManager->get(HiddenObjectHelper::class);
+        $hiddenObjectHelper->registerHiddenObjectInExtbaseSession(
+            $this->eventRepository,
+            $this->request,
+            'event'
+        );
         $this->addValidationForVideoLink();
         $this->applyLocationAsMandatoryIfNeeded();
         $this->arguments->getArgument('event')
@@ -228,7 +323,7 @@ class EventController extends AbstractController
      */
     public function activateAction(int $event): void
     {
-        $eventObject = $this->eventRepository->findHiddenEntry($event);
+        $eventObject = $this->eventRepository->findHiddenObject($event);
         $eventObject->setHidden(false);
         $this->eventRepository->update($eventObject);
 
@@ -269,5 +364,147 @@ class EventController extends AbstractController
         }
 
         return (bool)$this->mail->send();
+    }
+
+    /**
+     * Delete VideoLink if empty in Request.
+     * Extbase can not set deleted=1 itself.
+     *
+     * @param Event $event
+     */
+    protected function deleteVideoLinkIfEmpty(Event $event): void
+    {
+        if (
+            $event->getVideoLink() instanceof Link
+            && empty($event->getVideoLink()->getLink())
+        ) {
+            $this->linkRepository->remove($event->getVideoLink());
+            $event->setVideoLink(null);
+        }
+    }
+
+    /**
+     * Frontend insertion of events also has to respect location, if configured in ExtConf
+     */
+    protected function applyLocationAsMandatoryIfNeeded(): void
+    {
+        if ($this->extConf->getLocationIsRequired()) {
+            /** @var ValidatorResolver $validatorResolver */
+            $validatorResolver = $this->objectManager->get(ValidatorResolver::class);
+            $notEmptyValidator = $validatorResolver->createValidator(NotEmptyValidator::class);
+
+            /** @var ConjunctionValidator $eventValidator */
+            $eventValidator = $this->arguments->getArgument('event')->getValidator();
+            /** @var ConjunctionValidator $conjunctionValidator */
+            $conjunctionValidator = $eventValidator->getValidators()->current();
+            /** @var GenericObjectValidator $genericEventValidator */
+            $genericEventValidator = $conjunctionValidator->getValidators()->current();
+            $genericEventValidator->addPropertyValidator(
+                'location',
+                $notEmptyValidator
+            );
+        }
+    }
+
+    /**
+     * Add organizer.
+     *
+     * In a HTML-Template you can change the user uid if you want
+     * So it's better to add the organizer here in PHP
+     *
+     * @param string $argument
+     * @return bool
+     */
+    protected function addOrganizer(string $argument): bool
+    {
+        if ($this->request->hasArgument($argument)) {
+            $event = $this->request->getArgument($argument);
+            if (is_array($event) && array_key_exists('organizer', $event)) {
+                $organizerOfCurrentUser = $this->userRepository->getFieldFromUser('tx_events2_organizer');
+                if (MathUtility::canBeInterpretedAsInteger($organizerOfCurrentUser)) {
+                    $event['organizer'] = $organizerOfCurrentUser;
+                    // per default it is not allowed to add new Arguments manually. So we have to register them.
+                    // allow mapping of organizer
+                    $this->arguments
+                        ->getArgument($argument)
+                        ->getPropertyMappingConfiguration()
+                        ->allowProperties('organizer');
+                    // allow creation
+                    $this->arguments
+                        ->getArgument($argument)
+                        ->getPropertyMappingConfiguration()
+                        ->forProperty('organizer')
+                        ->setTypeConverterOption(
+                            'TYPO3\CMS\Extbase\Property\TypeConverter\PersistentObjectConverter',
+                            PersistentObjectConverter::CONFIGURATION_CREATION_ALLOWED,
+                            true
+                        )
+                        ->allowProperties('organizer');
+                    $this->request->setArgument($argument, $event);
+                } else {
+                    return false;
+                }
+            } else {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Files will be uploaded in typeConverter
+     * If newAction was called with a valid request something went wrong
+     * In that case all uploaded files have to be deleted
+     *
+     * @param string $argument
+     */
+    protected function deleteUploadedFilesOnValidationErrors(string $argument): void
+    {
+        if ($this->getControllerContext()->getRequest()->hasArgument($argument)) {
+            /** @var Event $event */
+            $event = $this->getControllerContext()->getRequest()->getArgument($argument);
+            if ($event instanceof Event) {
+                $images = $event->getImages();
+                foreach ($images as $image) {
+                    $orig = $image->getOriginalResource();
+                    $orig->delete();
+                }
+            }
+        }
+    }
+
+    /**
+     * Remove videoLink if empty
+     * Add special validation for VideoLink
+     * I can't add this validation to LinkModel, as such a validation would be also valid for organizer link.
+     */
+    protected function addValidationForVideoLink(): void
+    {
+        if (
+            $this->request->hasArgument('event') &&
+            $eventRaw = $this->request->getArgument('event') &&
+                empty($eventRaw['videoLink']['link'])
+        ) {
+            /** @var ValidatorInterface $regExpValidator */
+            $regExpValidator = GeneralUtility::makeInstance(RegularExpressionValidator::class, [
+                'regularExpression' => '~^(|http:|https:)//(|www.)youtube(.*?)(v=|embed/)([a-zA-Z0-9_-]+)~i',
+            ]);
+            /** @var ValidatorInterface $genericObjectValidator */
+            $genericObjectValidator = GeneralUtility::makeInstance(GenericObjectValidator::class);
+            $genericObjectValidator->addPropertyValidator('link', $regExpValidator);
+
+            // modify current validator of event
+            $event = $this->arguments->getArgument('event');
+            /** @var ConjunctionValidator $eventValidator */
+            $eventValidator = $event->getValidator();
+            $validators = $eventValidator->getValidators();
+            $validators->rewind();
+            $eventValidator = $validators->current();
+            $validators = $eventValidator->getValidators();
+            $validators->rewind();
+            /** @var GenericObjectValidator $eventValidator */
+            $eventValidator = $validators->current();
+            $eventValidator->addPropertyValidator('videoLink', $genericObjectValidator);
+        }
     }
 }
