+++ Classes/Controller/VideoController.php
@@ -14,8 +14,24 @@ namespace JWeiland\Events2\Controller;
 /*
  * A simple controller to show video-link as YouTube-Implementation
  */
+
+use JWeiland\Events2\Configuration\ExtConf;
+use JWeiland\Events2\Domain\Repository\EventRepository;
+use JWeiland\Events2\Domain\Repository\UserRepository;
+use TYPO3\CMS\Core\Mail\MailMessage;
+
 class VideoController extends AbstractController
 {
+    /**
+     * @var EventRepository
+     */
+    protected $eventRepository;
+
+    public function __construct(EventRepository $eventRepository)
+    {
+        $this->eventRepository = $eventRepository;
+    }
+
     /**
      * @param int $event
      */
