+++ Classes/Importer/ImporterInterface.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace JWeiland\Events2\Importer;
 
 use TYPO3\CMS\Core\Resource\FileInterface;
+use TYPO3\CMS\Scheduler\Task\AbstractTask;
 
 /*
  * An Interface as base for all event importer classes
@@ -19,12 +20,26 @@ use TYPO3\CMS\Core\Resource\FileInterface;
 interface ImporterInterface
 {
     /**
-     * Check, if File is valid for this importer
+     * Set task
+     * Needed, to get the StoragePid
+     *
+     * @param AbstractTask $task
+     */
+    public function setTask(AbstractTask $task): void;
+
+    /**
+     * Set file to import
      *
      * @param FileInterface $file
+     */
+    public function setFile(FileInterface $file): void;
+
+    /**
+     * Check, if File is valid for this importer
+     *
      * @return bool
      */
-    public function isValid(FileInterface $file): bool;
+    public function checkFile(): bool;
 
     /**
      * Import XML file
